Declare joining tolerance parameters as float to avoid compiler warning.
git-svn-id: https://svn.isais.edf.fr/mfee/saturne/trunk@2066 e44c5d27-f830-0410-8ce8-974f7a764be6

+++ users/base/cs_user_periodicity.c
@@ -100,9 +100,9 @@ BEGIN_C_DECLS
 void
 cs_user_periodicity(void)
 {
-  int     join_num;
-  int     verbosity = 1;
-  double  fraction = 0.10, plane = 0.25;
+  int    join_num;
+  int    verbosity = 1;
+  float  fraction = 0.10, plane = 0.25;
   return; /* REMOVE_LINE_FOR_USE_OF_SUBROUTINE */
 
   fraction = 0.10;
